fixed the 'Surprise me!' toggles
The checkbox and the counter were able to get out of sync, ie. the box could be unchecked but the counter would still be disabled. Now when the checkbox is clicked, it enables the counter if it's unchecked, and disables the counter if it's checked. Before, it would just switch the state of the counter