Implemented Way of the Mouse randomizer
If Way of the Mouse is one of the landscapes, then a randomizer for $2-$3 cards is revealed.
If a Liaison is chosen by that randomizer, then an Ally is added to the table if there isn't one already
If a Fate or Doom card is chosen, then a randomizer for Boons/Hexes is revealed.
There is also a space for associated States, Artifacts Heirlooms, and non-Supply Piles